pyproject.toml
[tool.poetry]
name = "spacelauncher"
version = "1.0.0"
description = "Space traffic and launch window tracking service"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
aiohttp = "^3.8.6"
poliastro = "^0.17.0"
astropy = "^5.3.4"
numpy = "^1.24.3"
pandas = "^2.1.1"
python-dotenv = "^1.0.0"
torch = "^2.1.0"
celery = "^5.3.4"
redis = "^5.0.1"
tenacity = "^8.2.3"
apscheduler = "^3.10.4"
structlog = "^23.2.0"
skyfield = "^1.46"
prometheus-client = "^0.17.1"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.21"
python-json-logger = "^2.0.7"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
black = "^23.9.1"
mypy = "^1.5.1"
bandit = "^1.7.5"
pytest-mock = "^3.11.1"
responses = "^0.23.3"
pytest-env = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=app --cov-report=html"
